fix(transfer): rename model file from transactions to transaction
